g620-a2: Fix #### network operator names
Huawei's modem is not pulling in network operator names for RIL.
The response to RIL_REQUEST_OPERATOR contains #### in place of the
network name. Rely on selective-spn-conf.xml which is provided by
vendor/cm as the primary source of network names. SPNs which do not
exist in this reference fall-back to the result from RIL.

Change-Id: If5896964ceba9c1c365a296c405eadf9a61b60bf